requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
openai>=1.30.1
requests==2.31.0
streamlit==1.34.0
audio-recorder-streamlit==0.0.8
